move Coin, CoinState and RespondToPhUpdates into chia-protocol
