+++ chia-protocol/Cargo.toml
@@ -9,13 +9,14 @@ homepage = "https://github.com/Chia-Network/chia_rs/chia-protocol/"
 repository = "https://github.com/Chia-Network/chia_rs/chia-protocol/"
 
 [features]
-py-bindings = ["dep:pyo3"]
+py-bindings = ["dep:pyo3", "dep:chia_py_streamable_macro"]
 
 [dependencies]
-pyo3 = { version = "=0.15.1", features = ["extension-module"], optional = true }
+pyo3 = { version = "=0.15.1", features = ["extension-module", "multiple-pymethods"], optional = true }
 sha2 = "=0.10.2"
 hex = "=0.4.3"
 chia_streamable_macro = { version = "=0.2.3", path = "../chia_streamable_macro" }
+chia_py_streamable_macro = { path = "../chia_py_streamable_macro", version = "=0.1.0", optional = true }
 clvmr = "=0.1.24"
 
 [lib]
