+++ chia-protocol/src/lib.rs
@@ -1,3 +1,10 @@
 pub mod bytes;
 pub mod chia_error;
+pub mod coin;
+pub mod coin_state;
+#[cfg(feature = "py-bindings")]
+pub mod from_json_dict;
+pub mod respond_to_ph_updates;
 pub mod streamable;
+#[cfg(feature = "py-bindings")]
+pub mod to_json_dict;
