+++ chia_py_streamable_macro/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "chia_py_streamable_macro"
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+description = "Derive macro to create python bindings for Chia types"
+authors = ["Arvid Norberg <arvid@libtorrent.org>"]
+homepage = "https://github.com/Chia-Network/chia_rs/chia_py_streamable_macro/"
+repository = "https://github.com/Chia-Network/chia_rs/chia_py_streamable_macro/"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "1.0.86"
+quote = "1.0.15"
