+++ chia_py_streamable_macro/src/lib.rs
@@ -65,7 +65,7 @@ pub fn py_streamable_macro(input: TokenStream) -> TokenStream {
                     #[staticmethod]
                     pub fn from_bytes(blob: &[u8]) -> PyResult<Self> {
                         let mut input = std::io::Cursor::<&[u8]>::new(blob);
-                        Self::parse(&mut input).map_err(|e| <chia_protocol::chia_error::Error as Into<PyErr>>::into(e))
+                        Self::parse(&mut input).map_err(|e| <chia_error::Error as Into<PyErr>>::into(e))
                     }
 
                     // returns the type as well as the number of bytes read from the buffer
@@ -78,7 +78,7 @@ pub fn py_streamable_macro(input: TokenStream) -> TokenStream {
                             std::slice::from_raw_parts(blob.buf_ptr() as *const u8, blob.len_bytes())
                         };
                         let mut input = std::io::Cursor::<&[u8]>::new(slice);
-                        Self::parse(&mut input).map_err(|e| <chia_protocol::chia_error::Error as Into<PyErr>>::into(e)).map(|v| (v, input.position() as u32))
+                        Self::parse(&mut input).map_err(|e| <chia_error::Error as Into<PyErr>>::into(e)).map(|v| (v, input.position() as u32))
                     }
 
                     pub fn get_hash<'p>(&self, py: Python<'p>) -> PyResult<&'p pyo3::types::PyBytes> {
@@ -88,7 +88,7 @@ pub fn py_streamable_macro(input: TokenStream) -> TokenStream {
                     }
                     pub fn to_bytes<'p>(&self, py: Python<'p>) -> PyResult<&'p pyo3::types::PyBytes> {
                         let mut writer = Vec::<u8>::new();
-                        self.stream(&mut writer).map_err(|e| <chia_protocol::chia_error::Error as Into<PyErr>>::into(e))?;
+                        self.stream(&mut writer).map_err(|e| <chia_error::Error as Into<PyErr>>::into(e))?;
                         Ok(pyo3::types::PyBytes::new(py, &writer))
                     }
