+++ wheel/Cargo.toml
@@ -19,6 +19,6 @@ chia = { path = "..", features = ["py-bindings"] }
 chia-protocol = { path = "../chia-protocol", features = ["py-bindings"]  }
 clvmr = "=0.1.24"
 pyo3 = { version = "=0.15.1", features = ["extension-module", "multiple-pymethods"] }
-py_streamable = { path = "py_streamable" }
+chia_py_streamable_macro = { path = "../chia_py_streamable_macro" }
 chia_streamable_macro = { version = "0.2.3", path = "../chia_streamable_macro" }
 hex = "=0.4.3"
