--- wheel/py_streamable/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "py_streamable"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = "1.0.86"
-quote = "1.0.15"
